refactor: extract asset search hook and dropdown
